requirements.txt
python-telegram-bot
requests
python-dotenv
tenacity
aiohttp
pytz